MySQL.pyを更新しました。

+++ models/MySQL.py
@@ -266,22 +266,26 @@ class MySQL:
             cursor = self.dbh.cursor()
             
             #IDの検索
-            stmt = "SELECT id, q_id FROM answer WHERE q_id = {}".format(question_id)
+            stmt = "SELECT id FROM answer WHERE q_id = {}".format(question_id)
             cursor.execute(stmt)
-            id_data = cursor.fetchall()
+            a_id = cursor.fetchall()
 
+            if id_data:
+                #user_answerで一致する回答IDの削除
+                stmt = "DELETE FROM user_answer WHERE a_id = {}".format()
+                cursor.execute(stmt)
+                
+                #回答の削除
+                stmt = "DELETE FROM answer WHERE q_id = {}".format(question_id)
+                cursor.execute(stmt)
+            
             #user_questionで一致する質問IDの削除
-
-
-            #回答の削除
-            stmt = "DELETE FROM answer WHERE q_id = {}".format(question_id)
+            stmt = "DELETE FROM user_question WHERE q_id = {}".format(question_id)
             cursor.execute(stmt)
 
             #質問の削除
             stmt = "DELETE FROM question WHERE id = {}".format(question_id)
             cursor.execute(stmt)
-
-            stmt = "DELETE FROM user_question WHERE q_id = "
         
         except mysql.connector.Error as err:
             print(err)#テスト用
@@ -489,7 +493,9 @@ class MySQL:
 
         try:
             self._open()
-            stmt = "SELECT question.id, question.title, question.category, questuon.regist_date FROM user_question LEFT JOIN question ON user_question.q_id = question.id WHERE user_question.user_id >= '{}'".format(user_id)
+            stmt = "SELECT question.id, question.title, question.category, questuon.regist_date FROM user_question \
+                LEFT JOIN question ON user_question.q_id = question.id \
+                WHERE user_question.user_id = '{}'".format(user_id)
             cursor = self.dbh.cursor()
             cursor.execute(stmt)
             user_questions = cursor.fetchall()
@@ -509,14 +515,14 @@ class MySQL:
     戻り値：データ配列（回答ID,日付,回答内容）
     機　能：ユーザがした質問の一覧を渡す
     """
-    def extract_user_question(self, user_id):
+    def extract_user_answer(self, user_id):
         user_answers = []
 
         try:
             self._open()
             stmt = "SELECT answer.id, answer.regist_date answer.text FROM extract_user_answer \
                 LEFT JOIN answer ON extract_user_answer.a_id = answer.id \
-                WHERE user_answer.user_id >= '{}'".format(user_id)
+                WHERE user_answer.user_id = '{}'".format(user_id)
             cursor = self.dbh.cursor()
             cursor.execute(stmt)
             user_answers = cursor.fetchall()
